+++ src/main/java/de/samply/core/ExporterParameters.java
@@ -14,5 +14,6 @@ public record ExporterParameters(
     String queryDescription,
     String queryContactId,
     String queryContext,
+    String queryExecutionContactId,
     LocalDate queryExpirationDate,
     Format outputFormat) {}