+++ src/main/resources/db/migration/V009__add_query_execution.sql
@@ -1,0 +1,2 @@
+ALTER TABLE samply.query_execution
+    ADD COLUMN contact_id TEXT;